requirements.txt
customtkinter>=5.2.0
python-vlc>=3.0.18121
sounddevice>=0.4.6
vosk>=0.3.45
pyttsx3>=2.90
rapidfuzz>=3.5.2
numpy>=1.24.0
pydub>=0.25.1
packaging
requests
pillow